Thu, Jun 29, 2023, 11:08 PM +02:00

+++ week04/pset4/filter-less/helpers.c
@@ -94,10 +94,10 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
             int redtest = round(blurredRed / divisor);
             int greentest = round(blurredGreen  / divisor);
             int bluetest = round(blurredBlue / divisor);
-            
-            image[i][j].rgbtRed = round(blurredRed / divisor);
-            image[i][j].rgbtGreen = round(blurredGreen  / divisor);
-            image[i][j].rgbtBlue = round(blurredBlue / divisor);
+
+            image[i][j].rgbtRed = round(blurredRed / (float) divisor);
+            image[i][j].rgbtGreen = round(blurredGreen  / (float) divisor);
+            image[i][j].rgbtBlue = round(blurredBlue / (float) divisor);
 
             // Reset colors
             blurredRed = 0;
